Completed Development Of The Note Interface

+++ middleware-common/src/main/java/com/jacolp/constant/AuditConstant.java
@@ -3,5 +3,5 @@ package com.jacolp.constant;
 public class AuditConstant {
     public static final Short REJECT = 2;
     public static final Short PASS = 1;
-    public static final int WAIT = 0;
+    public static final Short WAIT = 0;
 }
